+++ src/zepto.fullpage.js
@@ -1,5 +1,5 @@
 /*!
- * zepto.fullpage.js v0.3.1 (https://github.com/yanhaijing/zepto.fullpage)
+ * zepto.fullpage.js v0.5.0 (https://github.com/yanhaijing/zepto.fullpage)
  * API https://github.com/yanhaijing/zepto.fullpage/blob/master/doc/api.md
  * Copyright 2014 yanhaijing. All Rights Reserved
  * Licensed under MIT (https://github.com/yanhaijing/zepto.fullpage/blob/master/LICENSE)
@@ -8,9 +8,6 @@
     if (typeof $ === 'undefined') {
         throw new Error('zepto.fullpage\'s script requires Zepto');
     }
-    $(document).on('touchmove', function(e) {
-        e.preventDefault();
-    });
     var fullpage = null;
     var d = {
         page: '.page',
@@ -19,12 +16,17 @@
         loop: false,
         drag: false,
         dir: 'v',
+        der: 0.1,
         change: function(data) {},
         beforeChange: function(data) {},
         afterChange: function(data) {},
         orientationchange: function(orientation) {}
     };
 
+    function touchmove(e) {
+        e.preventDefault();
+    }
+    
     function fix(cur, pagesLength, loop) {
         if (cur < 0) {
             return !!loop ? pagesLength - 1 : 0;
@@ -39,8 +41,13 @@
     }
 
     function move($ele, dir, dist) {
-        var translate = dir === 'v' ? 'translateY' : 'translateX';
-        $ele.css({'-webkit-transform':translate + '(' + dist + 'px)','transform':translate + '(' + dist + 'px)'});
+        var xPx = "0px" , yPx = "0px";
+        if(dir === 'v') yPx = dist+"px";
+        else xPx = dist + "px";
+        $ele.css({
+            '-webkit-transform' : 'translate3d(' + xPx + ', ' + yPx + ', 0px);',
+            'transform' : 'translate3d(' + xPx + ', ' + yPx + ', 0px);'
+        });
     }
 
     function init(option) {
@@ -58,7 +65,7 @@
         that.pagesLength = that.$pages.length;
         that.update();
         that.initEvent();
-        that.status = 1;
+        that.start();
     }
 
     function Fullpage($this, option) {
@@ -100,8 +107,8 @@
                     return 0;
                 }
 
-                var sub = that.o.dir === 'v' ? e.changedTouches[0].pageY - that.startY : e.changedTouches[0].pageX - that.startX;
-                var der = (sub > 30 || sub < -30) ? sub > 0 ? -1 : 1 : 0;
+                var sub = that.o.dir === 'v' ? (e.changedTouches[0].pageY - that.startY)/that.height : (e.changedTouches[0].pageX - that.startX)/that.width;
+                var der = (sub > that.o.der || sub < -that.o.der) ? sub > 0 ? -1 : 1 : 0;
 
                 that.moveTo(that.curIndex + der, true);
             });
@@ -110,11 +117,15 @@
                     if (!that.status) {return 1;}
                     //e.preventDefault();
                     if (that.movingFlag) {
+                        that.startX = e.targetTouches[0].pageX;
+                        that.startY = e.targetTouches[0].pageY;
                         return 0;
                     }
 
                     var y = e.changedTouches[0].pageY - that.startY;
+                    if( (that.curIndex==0 && y>0) || (that.curIndex===that.pagesLength-1 && y<0) ) y/=2;
                     var x = e.changedTouches[0].pageX - that.startX;
+                    if( (that.curIndex==0 && x>0) || (that.curIndex===that.pagesLength-1 && x<0) ) x/=2;
                     var dist = (that.o.dir === 'v' ? (-that.curIndex * that.height + y) : (-that.curIndex * that.width + x));
                     $this.removeClass('anim');
                     move($this, that.o.dir, dist);
@@ -137,16 +148,25 @@
             }, false);
         },
 
+        holdTouch: function() {
+            $(document).on('touchmove', touchmove);
+        },
+        unholdTouch: function() {
+            $(document).off('touchmove', touchmove);
+        },
         start: function() {
             this.status = 1;
+            this.holdTouch();
         },
         stop: function() {
             this.status = 0;
+            this.unholdTouch();
         },
         moveTo: function(next, anim) {
             var that = this;
             var $this = that.$this;
             var cur = that.curIndex;
+
             next = fix(next, that.pagesLength, that.o.loop);
 
             if (anim) {
@@ -156,10 +176,15 @@
             }
 
             if (next !== cur) {
-                that.o.beforeChange({
+                var flag = that.o.beforeChange({
                     next: next,
                     cur: cur
                 });
+
+                // beforeChange 显示返回false 可阻止滚屏的发生
+                if (flag === false) {
+                    return 1;
+                }
             }
 
             that.movingFlag = true;
@@ -183,12 +208,17 @@
                     that.$pages.removeClass('cur').eq(next).addClass('cur');
                 }
             }, that.o.duration);
+
+            return 0;
         },
         movePrev: function(anim) {
             this.moveTo(this.curIndex - 1, anim);
         },
         moveNext: function(anim) {
             this.moveTo(this.curIndex + 1, anim);
+        },
+        getCurIndex: function () {
+            return this.curIndex;
         }
     });
 
@@ -198,14 +228,14 @@
         }
         return this;
     };
-    $.fn.fullpage.version = '0.3.1';
+    $.fn.fullpage.version = '0.5.0';
     //暴露方法
-    $.each(['update', 'moveTo', 'moveNext', 'movePrev', 'start', 'stop'], function(key, val) {
+    $.each(['update', 'moveTo', 'moveNext', 'movePrev', 'start', 'stop', 'getCurIndex', 'holdTouch', 'unholdTouch'], function(key, val) {
         $.fn.fullpage[val] = function() {
             if (!fullpage) {
                 return 0;
             }
-            fullpage[val].apply(fullpage, [].slice.call(arguments, 0));
+            return fullpage[val].apply(fullpage, [].slice.call(arguments, 0));
         };
     });
 }(Zepto, window));